Added a new ImageReader class, which uses ImageMagick++ to read image files, and produce openshot::Frame objects.

+++ include/FileReaderBase.h
@@ -28,7 +28,7 @@ namespace openshot
 		bool has_video;	///< Determines if this file has a video stream
 		bool has_audio;	///< Determines if this file has an audio stream
 		float duration;	///< Length of time (in seconds)
-		int file_size;	///< Size of file (in bytes)
+		long long file_size;	///< Size of file (in bytes)
 		int height;		///< The height of the video (in pixels)
 		int width;		///< The width of the video (in pixesl)
 		int pixel_format;	///< The pixel format (i.e. YUV420P, RGB24, etc...)
